+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### master / unreleased
 
+* Make syncDestroyed compatible with backbone >= 1.1.1
 * Fix #99: Restore compatibility with lodash; incompatible since 1.2.0 (Aleksandr Motsjonov)
 
 ### 1.3.0 / 2014-05-10